chore: Add temporary MySQL dump file for dev_db
This commit adds a temporary MySQL dump file (dev_db.sql) to the project, containing the current state of the development database. This file is intended to be used for backup and restore purposes during development.

The following changes were made:

- A new file (dev_db.sql) was added to the project root directory.
- The dev_db.sql file was generated using the 'mysqldump' command-line tool, using the '--all-databases' and '--single-transaction' options.

These changes were made to improve the development workflow by providing a simple and easy way to backup and restore the development database.

Note that this file should not be committed to the repository or used in production environments. It is intended for development use only.