feat: Add favicons to HTML documents
This commit adds a favicon to the website, which is displayed in the browser tab and bookmark bar to help identify the website and improve the user experience.

The following changes were made:

- A new favicon file was added to the project, in the 'image' directory.
- The HTML files in the project were updated to include a link to the favicon file in the 'head' section of the document.

These changes were made to improve the branding and user experience of the website, and to make it easier for users to identify and bookmark the site.

All affected HTML and image files have been updated and tested to ensure proper functionality.